crypto: reorg
